Add express production configs
Refs #48.
